move ttnn mocking and setup_mock_model_weights to mock_vllm_model.py
