+++ tests/mock_vllm_model.py
@@ -2,14 +2,23 @@
 #
 # SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
 
+import sys
 import copy
 import time
+import shutil
+from pathlib import Path
 from dataclasses import dataclass
 from typing import List
+from unittest.mock import MagicMock
 
 import torch
+from huggingface_hub import hf_hub_download
 
 from vllm.engine.metrics import logger
+
+# mock out ttnn fully so we can import ttnn without using it
+sys.modules["ttnn"] = MagicMock()
+sys.modules["ttnn.device"] = MagicMock()
 
 from models.demos.t3000.llama2_70b.tt.llama_common import (
     setup_llama_env,
@@ -21,6 +30,47 @@ from models.demos.t3000.llama2_70b.tt.llama_generation import (
 from models.demos.t3000.llama2_70b.tt.model_config import (
     get_model_config,
 )
+
+
+def setup_mock_model_weights(cache_root: str, weights_dir: str, hf_token: str):
+    if not hf_token:
+        raise ValueError(
+            "HuggingFace token (HF_TOKEN environment variable) is required"
+        )
+
+    metal_ckpt_dir = Path(weights_dir)
+    metal_tokenizer_path = metal_ckpt_dir / "tokenizer.model"
+    # Create path objects
+    # weights_dir = Path(cache_root) / "model_weights" / f"repacked-{model}"
+    metal_cache_path = (
+        Path(cache_root) / "tt_metal_cache" / f"cache_{metal_ckpt_dir.name}"
+    )
+
+    # Create directories
+    metal_ckpt_dir.mkdir(parents=True, exist_ok=True)
+    metal_cache_path.mkdir(parents=True, exist_ok=True)
+
+    # Define files to download
+    files_to_download = ["original/tokenizer.model", "original/params.json"]
+
+    # Download files using huggingface_hub
+    for file in files_to_download:
+        downloaded_path = hf_hub_download(
+            repo_id="meta-llama/Llama-3.1-70B-Instruct",
+            filename=file,
+            token=hf_token,
+            local_dir=metal_ckpt_dir,
+        )
+        # Move file to weights directory
+        target_path = metal_ckpt_dir / Path(file).name
+        shutil.move(downloaded_path, target_path)
+        print(f"Downloaded {file} to {target_path}")
+
+    # Clean up original directory if it exists and is empty
+    original_dir = metal_ckpt_dir / "original"
+    if original_dir.exists() and not any(original_dir.iterdir()):
+        original_dir.rmdir()
+    return str(metal_ckpt_dir), str(metal_tokenizer_path), str(metal_cache_path)
 
 
 def new_init_cache_enginer(self):
